refactor diffusion rotary_embedding
